# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  payment.OrderStatusResponse:
    properties:
      amount:
        type: number
      coupon_code:
        description: Код купона после успешной оплаты
        type: string
      currency:
        type: string
      message:
        type: string
      order_id:
        type: string
      size:
        type: string
      status:
        type: string
      style:
        type: string
      success:
        type: boolean
    type: object
  payment.PaymentNotificationRequest:
    properties:
      actionCode:
        description: Опционально для тестовой среды
        type: integer
      actionCodeDescription:
        description: Опционально для тестовой среды
        type: string
      amount:
        type: integer
      checksum:
        description: Дополнительные поля для валидации
        type: string
      currency:
        type: string
      date:
        type: integer
      ip:
        type: string
      orderDescription:
        type: string
      orderId:
        description: Опционально для тестовой среды
        type: string
      orderNumber:
        type: string
      orderStatus:
        description: Опционально для тестовой среды
        type: integer
    type: object
  payment.PurchaseCouponRequest:
    properties:
      domain:
        description: Домен партнера для White Label
        type: string
      email:
        type: string
      fail_url:
        type: string
      language:
        enum:
        - ru
        - en
        - es
        type: string
      return_url:
        type: string
      size:
        enum:
        - 21x30
        - 30x40
        - 40x40
        - 40x50
        - 40x60
        - 50x70
        type: string
      style:
        enum:
        - grayscale
        - skin_tone
        - pop_art
        - max_colors
        type: string
    required:
    - email
    - return_url
    - size
    - style
    type: object
  payment.PurchaseCouponResponse:
    properties:
      amont:
        type: string
      message:
        type: string
      order_id:
        type: string
      order_number:
        type: string
      payment_url:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для системы мозаичных купонов
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Mosaic API
  version: "1.0"
paths:
...
  /payment/notification:
    post:
      consumes:
      - application/json
      description: Handle payment webhook notification from payment gateway
      parameters:
      - description: Notification data
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/payment.PaymentNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Notification processed successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid notification data
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to process notification
          schema:
            additionalProperties: true
            type: object
      summary: Process payment notification
      tags:
      - payment
  /payment/options:
    get:
      description: Retrieve available payment methods and configurations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to get payment options
          schema:
            additionalProperties: true
            type: object
      summary: Get available payment options
      tags:
      - payment
  /payment/orders/{orderNumber}/status:
    get:
      description: Retrieve payment order status by order number
      parameters:
      - description: Order number
        in: path
        name: orderNumber
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payment.OrderStatusResponse'
        "400":
          description: Order number is required
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to get order status
          schema:
            additionalProperties: true
            type: object
      summary: Get order status
      tags:
      - payment
  /payment/purchase:
    post:
      consumes:
      - application/json
      description: Process coupon purchase and create payment order
      parameters:
      - description: Purchase request data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payment.PurchaseCouponRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payment.PurchaseCouponResponse'
        "400":
          description: Invalid request data
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to create order
          schema:
            additionalProperties: true
            type: object
      summary: Purchase coupon
      tags:
      - payment
  /payment/return:
    get:
      description: Handle payment return callback from payment gateway
      parameters:
      - description: Order number
        in: query
        name: orderNumber
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirect to success page
        "400":
          description: Order number is required
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to process payment return
          schema:
            additionalProperties: true
            type: object
      summary: Process payment return
      tags:
      - payment
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"